refactor: move ConnectivityService to dedicated service package for better code hygiene
