+++ src/main/kotlin/at/aau/hexabrawl/websocketserver/service/ConnectivityService.kt
@@ -1,5 +1,7 @@
-package at.aau.hexabrawl.websocketserver.model
+package at.aau.hexabrawl.websocketserver.service
 
+import at.aau.hexabrawl.websocketserver.model.GameState
+import at.aau.hexabrawl.websocketserver.model.UnitType
 import org.springframework.stereotype.Service
 
 /**
